refactor(portal): restructure admin management components and enhance vendor orders page
- Reorganize admin management components from `management/` to `managementpage/` directory
  - Move DeleteConfirmationModal, MessageModal, Pagination, UserDetailsModal,
    UserManagementHeader, and UserTable components
  - Update import paths in admin management page

- Completely refactor vendor orders page with modular component architecture
  - Extract functionality into separate components:
    - PageHeader: Handles page title and export functionality
    - SearchAndFilters: Manages search and filtering logic
    - OrderTable: Displays orders in tabular format
    - EmptyState: Shows when no orders are available
    - OrdersDetailModel: Modal for order details
    - Pagination: Handles pagination controls
    - ConfirmationModal: Confirmation dialogs
  - Improve code organization and maintainability
  - Reduce file size from 918 lines to more manageable components
  - Add proper TypeScript types and error handling
  - Implement better separation of concerns

- Total changes: 15 files modified, 1225 insertions, 716 deletions
